# packages/contracts/openapi/profile.v1.json
{
  "components": {
    "schemas": {
...
      "BusinessCreatedOut": {
        "properties": {
          "business": {
            "$ref": "#/components/schemas/BusinessOut"
          },
          "created": {
            "description": "False when the tenant already had this business",
            "title": "Created",
            "type": "boolean"
          },
          "onboarding": {
            "$ref": "#/components/schemas/OnboardingOut"
          },
          "prefill": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrefillOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the GSTIN filled in; null without one"
          }
        },
        "required": [
          "business",
          "created",
          "prefill",
          "onboarding"
        ],
        "title": "BusinessCreatedOut",
        "type": "object"
      },
      "BusinessIn": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "description": "First answers, stored with the business",
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "maxItems": 100,
            "title": "Answers",
            "type": "array"
          },
          "gstin": {
            "anyOf": [
              {
                "maxLength": 17,
                "minLength": 15,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "GSTIN, any case or spacing. Its PAN makes the business, and the registration is pre-filled from the GSTIN and the GSTIN lookup",
            "title": "Gstin"
          },
          "name": {
            "description": "The name the business goes by",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pan": {
            "anyOf": [
              {
                "maxLength": 12,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "PAN, any case or spacing; enough on its own when there is no GSTIN yet",
            "title": "Pan"
          },
          "registration_name": {
            "default": "",
            "description": "Name of the registration; the business name when empty",
            "maxLength": 200,
            "title": "Registration Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "BusinessIn",
        "type": "object"
      },
...
      "Page_BusinessSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BusinessSummaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Send as cursor to read the next page; null on the last page",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[BusinessSummaryOut]",
        "type": "object"
      },
...
      "Problem": {
        "description": "The body of every error response.",
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Correlation Id"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An access token from the identity service. Read when CW_AUTH_MODE is dual or token; required when it is token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "compliancewatch-profile",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/businesses": {
      "get": {
        "operationId": "list_businesses_v1_businesses_get",
        "parameters": [
          {
            "description": "Keeps the businesses whose name, PAN or GSTIN contains this, any case",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keeps the businesses whose name, PAN or GSTIN contains this, any case",
              "title": "Q"
            }
          },
          {
            "description": "Items per page, 1 to 200",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Items per page, 1 to 200",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "The next_cursor of the previous page; absent for the first page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The next_cursor of the previous page; absent for the first page",
              "title": "Cursor"
            }
          },
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BusinessSummaryOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The tenant's businesses by name, a page at a time",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      },
      "post": {
        "description": "A GSTIN makes the business from the PAN inside it and pre-fills the registration from the\nGSTIN lookup; a PAN alone makes the business with no registration. The answers are stored\nwith it, and the first onboarding question comes back. A business the tenant already has\n(the same PAN) is answered with ``created`` false.",
        "operationId": "create_business_v1_businesses_post",
        "parameters": [
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "A new value for each new request, such as a UUID; a retry sends the same value and gets the first response back for 24 hours",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "A new value for each new request, such as a UUID; a retry sends the same value and gets the first response back for 24 hours",
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[!-~]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a business from its GSTIN, or from its PAN alone",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      }
    },
    "/v1/businesses/{business_id}": {
      "get": {
        "operationId": "read_business_v1_businesses__business_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One business with its registrations and stored values",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      },
      "patch": {
        "description": "All or nothing: one value the ontology refuses stores none of the changes. Every node\nthat changed publishes ``profile.updated``, which recomputes the business's obligations.",
        "operationId": "update_business_v1_businesses__business_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessPatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Store answers across the business and its registrations, or rename it",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      }
    },
    "/v1/businesses/{business_id}/onboarding": {
      "get": {
        "operationId": "onboarding_v1_businesses__business_id__onboarding_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The next onboarding question with its wording and options, and the progress",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      }
    },
    "/v1/businesses/{business_id}/registrations": {
      "post": {
        "operationId": "add_registration_v1_businesses__business_id__registrations_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant UUID. A user's access token names the tenant, so a user leaves the header out or repeats that tenant; a service token names one here with the tenant:act scope.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "A new value for each new request, such as a UUID; a retry sends the same value and gets the first response back for 24 hours",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "A new value for each new request, such as a UUID; a retry sends the same value and gets the first response back for 24 hours",
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[!-~]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationAddIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a GSTIN registration to a business and pre-fill it",
        "tags": [
          "public",
          "businesses"
        ],
        "x-roles": [
          "owner",
          "staff",
          "ca_admin",
          "ca_staff",
          "compliance_lead"
        ]
      }
    },
...
  }
}